pyproject.toml
[tool.poetry]
name = "menobot"
version = "0.1.0"
description = "Python-Poetry template"
authors = ["Ledmington <ledmington.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.11"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
